Fixed the remaining warnings, added various NOLINT annotations

+++ libraries/data/include/mcrl2/data/detail/machine_word.h
@@ -177,7 +177,7 @@ std::size_t div_triple_doubleword(std::size_t n1, std::size_t n2, std::size_t n3
 
 inline std::size_t sqrt_word(const std::size_t n)
 {
-  return sqrt(n);
+  return static_cast<std::size_t>(sqrt(static_cast<double>(n)));
 }
 
 std::size_t sqrt_doubleword(std::size_t n1, std::size_t n2);
